+++ tests/test_main.py
@@ -358,6 +358,9 @@ class TestMain(TmpDirTestCase):
         assert_str_equalish(out, data)
 
 
+    ############################################################################
+    # Entrypoint
+
     def test_image_entrypoint(self):
         '''Verify scuba doesn't interfere with the configured image ENTRYPOINT'''
 
@@ -383,6 +386,101 @@ class TestMain(TmpDirTestCase):
         out, _ = self.run_scuba(['testalias'])
         assert_str_equalish('\n'.join(['success']*2), out)
 
+
+    def test_entrypoint_override(self):
+        '''Verify --entrypoint override works'''
+        with open('.scuba.yml', 'w') as f:
+            f.write('''
+                image: scuba/entrypoint-test
+                aliases:
+                  testalias:
+                    script:
+                      - echo $ENTRYPOINT_WORKS
+                ''')
+
+        test_str = 'This is output from the overridden entrypoint'
+
+        with open('new.sh', 'w') as f:
+            f.write('#!/bin/sh\n')
+            f.write('echo "{}"\n'.format(test_str))
+        make_executable('new.sh')
+
+        args = [
+            '--entrypoint', os.path.abspath('new.sh'),
+            'true',
+        ]
+        out, _ = self.run_scuba(args)
+        assert_str_equalish(test_str, out)
+
+
+    def test_entrypoint_override_none(self):
+        '''Verify --entrypoint override (to nothing) works'''
+        with open('.scuba.yml', 'w') as f:
+            f.write('''
+                image: scuba/entrypoint-test
+                aliases:
+                  testalias:
+                    script:
+                      - echo $ENTRYPOINT_WORKS
+                ''')
+
+        args = [
+            '--entrypoint', '',
+            'testalias',
+        ]
+        out, _ = self.run_scuba(args)
+
+        # Verify that ENTRYPOINT_WORKS wasn not set by the entrypoint
+        # (because it didn't run)
+        self.assertNotIn('success', out)
+
+
+    def test_yaml_entrypoint_override(self):
+        '''Verify entrypoint in .scuba.yml works'''
+        with open('.scuba.yml', 'w') as f:
+            f.write('''
+                image: scuba/entrypoint-test
+                entrypoint: "./new.sh"
+                ''')
+
+        test_str = 'This is output from the overridden entrypoint'
+
+        with open('new.sh', 'w') as f:
+            f.write('#!/bin/sh\n')
+            f.write('echo "{}"\n'.format(test_str))
+        make_executable('new.sh')
+
+        args = [
+            'true',
+        ]
+        out, _ = self.run_scuba(args)
+        assert_str_equalish(test_str, out)
+
+
+    def test_entrypoint_override_none(self):
+        '''Verify "none" entrypoint in .scuba.yml works'''
+        with open('.scuba.yml', 'w') as f:
+            f.write('''
+                image: scuba/entrypoint-test
+                entrypoint:
+                aliases:
+                  testalias:
+                    script:
+                      - echo $ENTRYPOINT_WORKS
+                ''')
+
+        args = [
+            'testalias',
+        ]
+        out, _ = self.run_scuba(args)
+
+        # Verify that ENTRYPOINT_WORKS wasn not set by the entrypoint
+        # (because it didn't run)
+        self.assertNotIn('success', out)
+
+
+    ############################################################################
+    # Image override
 
     def test_image_override(self):
         '''Verify --image works'''